+++ src/pages/dashboard/pages/Users.jsx
@@ -43,7 +43,7 @@ export function Users() {
         <thead className="border border-gold">
           <tr className="goldGradient text-base uppercase">
             {colsTitles.map((t, index) => (
-              <th className="px-6 py-2" key={index}>
+              <th className="border-r-2 border-r-blue px-6 py-2 last:border-none" key={index}>
                 {t}
               </th>
             ))}
