+++ src/pages/dashboard/pages/index.js
@@ -4,3 +4,4 @@ export { Blog } from "./Blog";
 export { Users } from "./Users";
 export { TableRow } from "./Utils";
 export { MoreActions } from "./MoreActions";
+export { UsdPriceManager } from "./UsdPriceManager";
